refactor: change costThreshold type from int to float in Benchmark class

+++ include/Benchmark.h
@@ -11,11 +11,11 @@ class Benchmark {
     // * Esses atributos são específicos do Benchmark, pois podem variar a cada execução
     double a, b, c;
     int seed;
-    int costThreshold;
+    float costThreshold;
 
   public:
     //@ Construtor que recebe os coeficientes e parametros do benchmark
-    Benchmark(double a, double b, double c, int seed, int costThreshold);
+    Benchmark(double a, double b, double c, int seed, float costThreshold);
 
     void run(Vector& V);
 };
